Makefile: use git submodule sync to ensure the URLs match .gitmodules
Otherwise, git prefers "config" to "gitmodules", which means the
autobuild repository's versions (which are global) stick around.

This workaround leaves a brief race condition in place, where the global
config file contains URLs for submodules. Because they are used
immediately afterwards by this script, it's only a brief window of
susceptibility (and it requires a git-submodule clash in order to
trigger it.)